+++ frontend/src/api/types/clients.ts
@@ -1,4 +1,5 @@
 import type { JwkKeyPairAlg } from './oidc';
+import type { JsonValue } from '$utils/jsonValue';
 
 export const AuthFlowDeviceCode = 'urn:ietf:params:oauth:grant-type:device_code';
 export type AuthFlow =
@@ -66,6 +67,9 @@ export interface UpdateClientRequest {
     backchannel_logout_uri?: string;
     /// Validation: PATTERN_GROUP
     restrict_group_prefix?: string;
+    /// Validation: JSON object, max 1024 serialized characters
+    claims?: JsonValue;
+    claims_at_root?: boolean;
     scim?: ScimClientRequestResponse;
 }
 
@@ -95,6 +99,8 @@ export interface ClientResponse {
     contacts?: string[];
     backchannel_logout_uri?: string;
     restrict_group_prefix?: string;
+    claims?: JsonValue;
+    claims_at_root: boolean;
     scim?: ScimClientRequestResponse;
 }